kernel: Add size parameter to K_{KERNEL,THREAD}_STACK_EXTERN
This commit adds the stack size parameter to the kernel
`K_KERNEL_STACK_EXTERN` and `K_THREAD_STACK_EXTERN` macros so that the
the `sizeof` operator can be used with the extern-ed stack symbols.

This aligns the behaviour of the `K_{KERNEL,THREAD}_STACK_EXTERN`
macros with that of the `K_{KERNEL,THREAD}_STACK_ARRAY_EXTERN` macros,
which also accept the `size` parameter.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>